requirements.txt
jupyterlab
Pillow
altair
pandas
numpy
scipy
matplotlib
docopt
